* feat: 엔티티에서 사용되는 원시값 포장객체와 일급컬렉션 생성
* refactor: 원시값 포장객체, 일급컬렉션 생성으로 관련 코드 리팩터링

- 기존 원시값 참조하는 로직 수정
- 테스트 코드 리팩터링
- 비지니스 로직 리팩터링

+++ src/main/java/kitchenpos/order/dto/OrderLineItemResponse.java
@@ -18,7 +18,7 @@ public class OrderLineItemResponse {
     }
 
     public static OrderLineItemResponse of(OrderLineItem orderLineItem) {
-        return new OrderLineItemResponse(orderLineItem.getSeq(), MenuResponse.of(orderLineItem.getMenu()), orderLineItem.getQuantity());
+        return new OrderLineItemResponse(orderLineItem.getSeq(), MenuResponse.of(orderLineItem.getMenu()), orderLineItem.getQuantity().toLong());
     }
 
     public Long getSeq() {
